backend/requirements.txt
fastapi[all]==0.135.1
alembic 
psycopg2-binary
celery
redis
aiohttp
requests
passlib[argon2]==1.7.4
pytest
PyJWT
aioboto3
